employee/requirement.txt
asgiref==3.7.2
Django==4.2.5
django-cors-headers==4.2.0
django-restframework==0.0.1
djangorestframework==3.14.0
Pillow==10.0.1
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
